DineroObject base class
