+++ dinero/card.py
@@ -1,6 +1,7 @@
 from dinero.log import log
+from dinero.base import DineroObject
 
-class CreditCard(object):
+class CreditCard(DineroObject):
     """
     A Customer resource. `Customer.create` uses the gateway to create a
     customer.  You can use this Customer object in calls to
@@ -10,6 +11,7 @@ class CreditCard(object):
     def __init__(self, customer_id, **kwargs):
         self.customer_id = customer_id
         self.data = kwargs
+
     @log
     def save(self):
         raise NotImplemented
@@ -18,23 +20,11 @@ class CreditCard(object):
     def delete(self):
         raise NotImplemented
 
-    def to_dict(self):
-        return vars(self)
-
-    def __getattr__(self, attr):
-        if attr == '__setstate__':
-            raise AttributeError
-        try:
-            return self.data[attr]
-        except KeyError as e:
-            raise AttributeError(e)
-
     def __setattr__(self, attr, val):
         if attr in ['customer_id', 'data']:
             self.__dict__[attr] = val
         else:
             self.data[attr] = val
-
 
     @classmethod
     def from_dict(cls, dict):
