+++ dinero/customer.py
@@ -2,9 +2,10 @@ from dinero import get_gateway
 from dinero.exceptions import InvalidCustomerException
 from dinero.log import log
 from dinero.card import CreditCard
+from dinero.base import DineroObject
 
 
-class Customer(object):
+class Customer(DineroObject):
     """
     A Customer resource. `Customer.create` uses the gateway to create a
     customer.  You can use this Customer object in calls to
@@ -59,17 +60,6 @@ class Customer(object):
         gateway = get_gateway(gateway_name)
         resp = gateway.add_card_to_customer(self, options)
         return CreditCard(customer=self, **resp)
-
-    def to_dict(self):
-        return vars(self)
-
-    def __getattr__(self, attr):
-        if attr == '__setstate__':
-            raise AttributeError
-        try:
-            return self.data[attr]
-        except KeyError as e:
-            raise AttributeError(e)
 
     def __setattr__(self, attr, val):
         if attr in ['gateway_name', 'customer_id', 'data']:
